book stat DSL로 변환

+++ src/main/kotlin/com/quid/kopring/book/model/response/BookStat.kt
@@ -2,6 +2,6 @@ package com.quid.kopring.book.model.response
 
 import com.quid.kopring.book.model.type.BookType
 
-data class BookStat(val type: BookType, val count: Int) {
+data class BookStat(val type: BookType, val count: Long) {
 
 }
